Separate event dispatch and event registration.
